requirements.txt
numpy
matplotlib
Adafruit-BNO055
RPi.GPIO-def
pynetworktables